// frontend/aada_web/src/api/generated/models/credentialRead.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * AADA LMS API
 * OpenAPI spec version: 1.0
 */

export interface CredentialRead {
  user_id: string;
  program_id: string;
  /** @maxLength 50 */
  credential_type: string;
  /** @maxLength 100 */
  cert_serial: string;
  issued_at: string;
  id: string;
}