// frontend/aada_web/src/api/generated/models/transcriptRead.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * AADA LMS API
 * OpenAPI spec version: 1.0
 */
import type { TranscriptReadGpa } from './transcriptReadGpa';
import type { TranscriptReadPdfUrl } from './transcriptReadPdfUrl';
import type { ModuleResult } from './moduleResult';

export interface TranscriptRead {
  id: string;
  user_id: string;
  program_id: string;
  gpa?: TranscriptReadGpa;
  generated_at: string;
  pdf_url?: TranscriptReadPdfUrl;
  modules?: ModuleResult[];
}